Moved the StoryPage to its own dart file
